+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project uses [semantic versioning](http://semver.org/).
 
+## master
+
+### Added
+
+- Added reset command.
+
 ## 0.10.1 - 2016-03-28
 
 ### Added